finished sync_members()
made two set of tuples.
first checks if sets are equal

Then use difference_update() to remove users who are neither new or needs updating data.

For new users we look for userIDs not present in database and if not found adds the new user to db.

after we take the remaining users and update them